PORTFOLIO
In this lab we developed a simple portfolio website using HTML and CSS. The main objective was to understand how a basic website is created and how different sections are connected on a single page. This lab helped us understand the overall structure of a webpage and how content is placed in an organized way. In this lab, we learn how to use HTML tags properly for headings, paragraphs, images, tables, lists and forms creating different sections like home, about, project and contact that help us understand how website are divided into meaningful parts. We also learn how  navigation links help users move easily from one section to another.  We practiced using CSS to improve the design of the webpage. we worked with colors, spacing, borders,  buttons and layouts. Styling the navigation bar, table and contact form helped me understand how CSS controls the look and feel of a website. We faced several problems such as alignment issues, spacing errors and styling mistakes. At first some parts of the piece did not look correct but fixing those errors helped us learn more clearly. We also learn the importance of writing clean and readable code. Keeping the HTML structure simple and using CSS properly made the website easier to manage. This portfolio showed us how even small changes in code can make a big difference in the final output. We gained knowledge in basic web development skills although this website is simple and made only for practice it helped us understand the fundamentals of HTML and CSS.